Move selector rules into parser
